% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{add_page_to_unit}
\alias{add_page_to_unit}
\title{Add an (existing) page to a unit}
\usage{
add_page_to_unit(module_id, unit_search_term, page_search_term)
}
\arguments{
\item{module_id}{Canvas module ID number}

\item{unit_search_term}{String to match the unit name - MUST match case!}

\item{page_search_term}{String to match the page name - does not need to match case}
}
\value{
Response from Canvas
}
\description{
Add an (existing) page to a unit
}
