% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quizzes.R
\name{add_quiz_groups_info}
\alias{add_quiz_groups_info}
\title{Add info about quiz groups to question info}
\usage{
add_quiz_groups_info(module_id, quizzes_info)
}
\arguments{
\item{module_id}{Canvas module ID number}

\item{quizzes_info}{A tibble of quiz question info minimally containing
variables \code{quiz_id} and \code{quiz_group_id}.}
}
\value{
The same \code{quizzes_info} tibble but with added details about the quiz
group each question belongs to - see \code{\link[=get_quiz_groups]{get_quiz_groups()}}
}
\description{
Given a tibble of question info, will retrieve the information about quiz
group info for each question and merge into the table. Great for creating
question banks!
}
