% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submissions_marks.R
\name{calc_quiz_marks}
\alias{calc_quiz_marks}
\title{Calculate "best-of" worksheet marks}
\usage{
calc_quiz_marks(
  module_code,
  academic_year,
  quiz_title = "Worksheet",
  quiz_data,
  max_score
)
}
\arguments{
\item{module_code}{A Sussex module code, e.g. "C8891".}

\item{academic_year}{A string containing the academic year for the desired
module as e.g. "22/23"; defaults to the current year as calculated by
[cnvs::get_ac_year()]. Does not need to be specified for
ongoing/non-academic modules}

\item{quiz_title}{Common element of the names of all quizzes to be included.
Defaults to "Worksheet". Passed to [cnvs::get_quiz_marks()].}

\item{quiz_data}{A long-form dataset of quiz scores, including `cand_no`
(candidate number), `week` (week of term), and `score` (quiz score). Use
this option if you have previously needed to download the quiz data and
make manual changes before calculating the quiz mark. If this argument is
empty, the function will instead download the quiz marks directly from
Canvas using the information in the following arguments.}

\item{max_score}{Maximum possible score for the quizzes}
}
\value{
A list with three elements: $all_marks, a tibble of all quiz scores for all
  students; $mean_marks, a tibble of final quiz mark for all students; and $dist, a histogram of the distribution of those final marks
}
\description{
Calculate "best-of" worksheet marks
}
