% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{canvas_setup}
\alias{canvas_setup}
\title{Setup Canvas for API use}
\usage{
canvas_setup(domain = "https://canvas.sussex.ac.uk")
}
\arguments{
\item{domain}{A text string URL of the homepage of your university's Canvas
site. Default is for University of Sussex.}
}
\value{
Sets the Canvas token and domain.
}
\description{
Wrapper for two `rcanvas` functions: [rcanvas::set_canvas_token()] and
[rcanvas::set_canvas_domain()]. Calls [Sys.getenv()] for setting the token,
so this must be set up ahead of time (see Details).
}
\details{
To use this function, you must obtain a Canvas token and set it as an
environmental variable.

To obtain a Canvas token:
  * Log into Canvas.
  * On the left bar, click Account, then Settings.
  * Click New Access Token and follow the instructions.

Save your Canvas token somewhere safe, where you can find it again. Then,
open your Renviron file (by using e.g. [usethis::edit_r_environ()]) and
create a new variable called  called `CANVAS_TOKEN` with your Canvas token in
quotes, making sure to keep the last line of the file empty. Save the file,
close it, and restart R (Session > Restart R) for the changes to take effect.
}
