% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnvs-package.R
\docType{package}
\name{cnvs-package}
\alias{cnvs-package}
\title{cnvs: 'Canvas LMS' 'API'}
\description{
Minimal client to access the 'Canvas LMS' 'API'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://cwickham.github.io/cnvs}
  \item \url{https://github.com/cwickham/cnvs}
  \item Report bugs at \url{https://github.com/cwickham/cnvs/issues}
}

}
\author{
\strong{Maintainer}: Charlotte Wickham \email{cwickham@gmail.com} [contributor]

Authors:
\itemize{
  \item Gábor Csárdi (gh package)
  \item Jennifer Bryan (gh package)
  \item Hadley Wickham (gh package)
}

}
\keyword{internal}
