% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnvs_token.R
\name{cnvs_token}
\alias{cnvs_token}
\alias{cnvs_token_exists}
\alias{cnvs_domain}
\title{Return the local user's Canvas token and domain}
\usage{
cnvs_token()

cnvs_token_exists()

cnvs_domain()
}
\value{
A string. For \code{cnvs_token()}, the return value has an S3 class
to ensure that simple printing strategies don't reveal the entire token.
Both functions error if the corresponding environment variable is not set.
}
\description{
\code{cnvs_token()} returns the Canvas API token from the \code{CANVAS_API_TOKEN}
environment variable. Most Canvas API requests require a token to prove
the request is authorized by a specific Canvas user.

\code{cnvs_domain()} returns the Canvas domain URL from the \code{CANVAS_DOMAIN}
environment variable. This is your institution's Canvas URL, e.g.
\verb{https://canvas.instructure.com} or \verb{https://myuni.instructure.com}.

You can generate a token from your Canvas profile settings page at
\code{{CANVAS_DOMAIN}/profile/settings} under "Approved Integrations".

See \code{\link[=cnvs_whoami]{cnvs_whoami()}} for more details on setting up your credentials.
}
\examples{
\dontrun{
cnvs_token()
cnvs_domain()

format(cnvs_token())
}
}
