% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload.R
\name{cnvs_upload}
\alias{cnvs_upload}
\title{Upload files to Canvas}
\usage{
cnvs_upload(endpoint, path, ...)
}
\arguments{
\item{endpoint}{Character string of the API endpoint for upload.  This depends on the required
context of file. For example, uploading a file to a course uses the endpoint
\code{"/api/v1/courses/:course_id/files"}.  Uploading a file as an assignment submission
has the endpoint \code{"/api/v1/courses/:course_id/assignments/:assignment_id/submissions/:user_id/files"}}

\item{path}{Path to file to upload}

\item{...}{Other parameters passed along to \code{\link{cnvs}()}.  Use these
to specify endpoint parameters, e.g. \code{course_id}, parameters for the file
upload, e.g. \code{parent_folder_path}, or other parameters to \code{\link{cnvs}()}
like \code{.token} or \code{.api_endpoint}.}
}
\description{
Upload a local file to a location on Canvas LMS. Read more about possible parameters
at \url{https://canvas.instructure.com/doc/api/file.file_uploads.html}.
}
\examples{
\dontrun{
## Upload a course file, ends up in "unfiled"
cnvs_upload("/api/v1/courses/:course_id/files",
  path = "notes.pdf", course_id = "1732420")

## Upload a course file to a specific folder
cnvs_upload("POST /api/v1/courses/:course_id/files",
  path = "notes.pdf", course_id = "1732420", parent_folder_path = "cnvs_files")
}
}
