% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnvs_whoami.R
\name{cnvs_whoami}
\alias{cnvs_whoami}
\title{Info on current Canvas user, domain and token}
\usage{
cnvs_whoami(.token = NULL, .api_url = NULL, .send_headers = NULL)
}
\arguments{
\item{.token}{Authentication token. Defaults to \code{\link[=cnvs_token]{cnvs_token()}}.}

\item{.api_url}{Canvas domain URL. Used if \code{endpoint} just contains a path.
Defaults to \code{CANVAS_DOMAIN} environment variable if set.}

\item{.send_headers}{Named character vector of header field values
(except \code{Authorization}, which is handled via \code{.token}). This can be
used to override or augment the default \code{User-Agent} header:
\code{"https://github.com/cwickham/cnvs"}.}
}
\value{
A \code{cnvs_response} object, which is also a \code{list}.
}
\description{
Reports name and Canvas login ID for the current authenticated user, the
Canvas domain and the first bit of the token.
}
\details{
Your Canvas domain is the address of your organization's Canvas site, e.g.
\verb{https://oregonstate.instructure.com/}. Set this via the \code{CANVAS_DOMAIN}
environment variable or pass it to \code{.api_url}.

Get a personal access token for your Canvas domain from
\code{{CANVAS_DOMAIN}/profile/settings}. The token itself is a string of
letters and digits. You can store it any way you like and provide explicitly
via the \code{.token} argument to \code{\link[=cnvs]{cnvs()}}.

However, many prefer to define environment variables \code{CANVAS_API_TOKEN}
and \code{CANVAS_DOMAIN} with these values in their \code{.Renviron} file. Add
lines that look like these, substituting your domain and token:

\if{html}{\out{<div class="sourceCode">}}\preformatted{CANVAS_DOMAIN="https://canvas.instructure.com"
CANVAS_API_TOKEN="your_token_here"
}\if{html}{\out{</div>}}

Put a line break at the end! If you're using an editor that shows line
numbers, there should be (at least) three lines, where the third one is empty.
Restart R for this to take effect. Call \code{cnvs_whoami()} to confirm
success.

To get complete information on the authenticated user, call
\code{cnvs("/api/v1/users/self")}.
}
\examples{
\dontrun{
cnvs_whoami()

## explicit token + domain
cnvs_whoami(
  .token = "your_token_here",
  .api_url = "https://canvas.instructure.com"
)
}
}
