% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{create_bullet_link}
\alias{create_bullet_link}
\title{Generate a linked bullet point}
\usage{
create_bullet_link(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[=create_link]{create_link()}}}
}
\value{
A single text string in the format "- \verb{linked text}" with two following newlines
}
\description{
Generate a linked bullet point
}
