% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{create_posit_link}
\alias{create_posit_link}
\title{Generate a link to a Posit list}
\usage{
create_posit_link(list_number, workspace_number)
}
\arguments{
\item{list_number}{The numerical ID of the list}

\item{workspace_number}{The numerical ID of the workspace that the list exists in}
}
\value{
A URL to a list on a Posit workspace as a text string
}
\description{
Generate a link to a Posit list
}
