% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quizzes.R
\name{create_quiz}
\alias{create_quiz}
\title{Create a Canvas quiz}
\usage{
create_quiz(module_id, title, ...)
}
\arguments{
\item{module_id}{Canvas module ID number}

\item{title}{Title of the quiz.}

\item{...}{Any number of additional arguments. These MUST correspond to valid
parameter names for an API request, as they will be automatically interpreted
as such. For instance, to set the parameter \code{quiz[quiz_type]}, provide the
argument \code{quiz_type = "assignment"} (or whichever allowed value you want.)
For a complete list of parameters and allowed values, see \link{the Quizzes API}
(https://developerdocs.instructure.com/services/canvas/resources/quizzes#method.quizzes/quizzes_api.create).}
}
\value{
Response from Canvas
}
\description{
An open-ended function with (extremely) few assumptions about the settings
the quiz should have. If you want an easier way to batch-create multiple
quizzes that DOES make (quite a few) more assumptions, see
\code{\link[=create_quizzes]{create_quizzes()}}.
}
\details{
The only essential information to provide is the quiz title, but I do
recommend providing more than this, unless you're quite happy with the
defaults (unlikely)!
}
