% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quizzes.R
\name{create_quiz_desc}
\alias{create_quiz_desc}
\title{Create quiz descriptions}
\usage{
create_quiz_desc(n_questions = 5, standard_time_limit = 12, challengr = FALSE)
}
\arguments{
\item{n_questions}{Total number of questions the quiz will contain}

\item{standard_time_limit}{Time limit in minutes for the "standard" amount of
time, i.e. before any extra time reasonable adjustments are implemented.
The function will automatically calculate and add info about 25\% and 50\%
extra time based on this number.}

\item{challengr}{Are the quizzes being created ChallengRs?}
}
\value{
A list of two descriptions, "practice" and "marked".
}
\description{
Create quiz descriptions
}
