% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quizzes.R
\name{create_quiz_groups}
\alias{create_quiz_groups}
\title{#' Create information about quiz groups
#'
#'}
\usage{
create_quiz_groups(module_id, quiz_id, args_data, ...)
}
\arguments{
\item{module_id}{Canvas module ID}

\item{quiz_id}{Canvas quiz ID}

\item{args_data}{Tibble of arguments, with variable names the same as the
parameter options (listed above) and values for each group in each row}

\item{...}{Parameter options (listed above) given as named arguments, e.g.
\code{name = "Hard Questions"}.}
}
\value{
Response from Canvas
}
\description{
which_week <- 1:11
}
\details{
module_id <- try(convenrhelpr::get_module_id(module_code), silent = TRUE)

if(inherits(module_id, "try-error")){
convenrhelpr::canvas_setup()
module_id <- convenrhelpr::get_module_id(module_code)
}

assignments <- rcanvas::get_assignment_list(module_id) |> tibble::as_tibble()

quiz_id <- assignments |>
dplyr::filter(name \%in\% paste("Week", which_week, "Worksheet", sep = " ")) |>
dplyr::pull(quiz_id)

groups_info <- tidyr::crossing(quiz_id,
group_name = c("Concepts", "Code", "Interpretation")) |>
dplyr::mutate(
pick_count = ifelse(group_name == "Interpretation", 3, 2),
points = 1,
module_id = module_id
)

Create groups in a quiz

Generate question groups in an existing quiz. There are only four valid
parameters that can be provided:
\itemize{
\item \code{name}: the name of the group
\item \code{pick_count}: how many questions should be randomly selected from the group
\item \code{question_points}: how many points each question is worth
\item \code{assessment_question_bank_id}: ID number of the assessment question bank to
link to. Unclear at the moment if/how this ID number can be accessed via the
API; otherwise, find it here:
https://canvas.your_uni.ac.uk/courses/module_id/question_banks/THIS_NUMBER
}

Parameters can be provided as either a tibble (for multiple groups at once)
or as named arguments. Either method will accept any combination of
parameters, and the argument method will accept none.
}
