% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quizzes.R
\name{create_quiz_questions}
\alias{create_quiz_questions}
\title{Post quiz questions to a Canvas quiz}
\usage{
create_quiz_questions(module_id, quiz_id, data)
}
\arguments{
\item{module_id}{Canvas module ID number}

\item{quiz_id}{Numerical Canvas ID number for the quiz to which the questions will be added.}

\item{data}{A tibble of quiz information with answer information nested by
row; for a template, use \link{quiz_question_example}.}
}
\value{
Response from Canvas
}
\description{
In order to create a question successfully, the variable names in \code{data}
MUST match the corresponding fields for the QuizQuestion and Answer
objects \href{https://developerdocs.instructure.com/services/canvas/resources/quiz_questions}{see Canvas API help documentation for options}.
Furthermore, the answer information must be held in its own tibble, nested
in the question data tibble, in a variable called \code{answers}. For a template,
use \link{quiz_question_example}.
}
