% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quizzes.R
\name{create_quizzes}
\alias{create_quizzes}
\title{Create (multiple) Canvas quizzes}
\usage{
create_quizzes(
  week,
  title,
  unlock_at,
  lock_at,
  module_id,
  description,
  quiz_type,
  time_limit,
  allowed_attempts,
  access_code
)
}
\arguments{
\item{week}{Week of term}

\item{title}{Title of the quiz to be created}

\item{unlock_at}{A datetime using \code{\link[=format_ISO8601]{format_ISO8601()}} indicating when the quiz
should be unlocked.}

\item{lock_at}{A datetime using \code{\link[=format_ISO8601]{format_ISO8601()}} indicating when the quiz
should be locked.}

\item{module_id}{Canvas module ID number}

\item{description}{Single string containing the text description for the quiz.
Accepts basic HTML formatting tags.}

\item{quiz_type}{Either one of "assignment", "practice_quiz",
"graded_survey", or "survey" (which will be applied to all) or a vector of
combinations of the same in the order in which they should be applied.}

\item{time_limit}{Time limit for the quiz in minutes.}

\item{allowed_attempts}{A number, with -1 corresponded to infinite attempts.}

\item{access_code}{Logical. If TRUE, a random 6-digit code will be generated.
If FALSE, no code will be required to access the quiz.}
}
\value{
Answer from Canvas; quizzes should appear on Canvas.
}
\description{
Create weekly worksheet quizzes on your Canvas site. For making multiple
quizzes at once, use in conjunction with \code{\link[purrr:pmap]{purrr::pmap()}} and a tibble of
associated quiz info generated by \code{\link[=create_quiz_info]{create_quiz_info()}}.
}
\details{
This function is built to work with others for generating multiple similar
quizzes at once. If you want a more basic version with fewer assumptions,
see \code{\link[=create_quiz]{create_quiz()}}.
}
\seealso{
\code{\link[=get_module_id]{get_module_id()}}
}
