% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sections.R
\name{create_section}
\alias{create_section}
\title{Create a new section in a module}
\usage{
create_section(module_id, section_name)
}
\arguments{
\item{module_id}{Canvas module ID number}

\item{section_name}{Name to give the new section}
}
\value{
A response from Canvas.
}
\description{
Create a new section in a module
}
