% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{create_term_info}
\alias{create_term_info}
\title{Create term information}
\usage{
create_term_info(
  term_start_date,
  term_end_date,
  has_break_weeks = TRUE,
  break_weeks = 11:13
)
}
\arguments{
\item{term_start_date}{First day of term (the first Monday)}

\item{term_end_date}{Last day of term (the last Friday)}

\item{has_break_weeks}{Logical. Does the module have break weeks (e.g. spring break)?}

\item{break_weeks}{Numerical vector indicating weeks out of the total number
of weeks that the term spans are break weeks. E.g. for an 11-week term, with three weeks of spring break, the
total number of weeks is 14: 10 weeks of term, three weeks break, and one
final week of term. In this case the input would be 11:13.}
}
\value{
A tibble of weeks and start and end dates
}
\description{
Make a tibble of information about term dates
}
