% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sections.R
\name{enroll_in_section}
\alias{enroll_in_section}
\title{Enroll a user in an (existing) section}
\usage{
enroll_in_section(section_id, user_id)
}
\arguments{
\item{section_id}{ID of the section to enroll the user in}

\item{user_id}{ID of the user}
}
\value{
Response from Canvas
}
\description{
Enroll a user in an (existing) section
}
