% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{find_weekday}
\alias{find_weekday}
\title{Calculate the date of the next (or previous) weekday from a date}
\usage{
find_weekday(date, weekday, direction = c("previous", "next"))
}
\arguments{
\item{date}{A datetime}

\item{weekday}{A day of the week as a number, with 1 = Sunday and 7 = Saturday, OR as a string with a weekday name}

\item{direction}{Either "previous" or "next"}
}
\value{
The date of the next \link{weekday} before or after the given \link{date}
}
\description{
Adapted from a \href{https://stackoverflow.com/questions/57893554/get-next-wednesday-date-after-a-date-with-r}{stack overflow solution by TimTeaFan}
}
