% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_ac_year}
\alias{get_ac_year}
\title{Create a string for the current academic year}
\usage{
get_ac_year(switch_date = as.Date(paste0(format(Sys.Date(), "\%Y"), "-09-01")))
}
\arguments{
\item{switch_date}{A date object containing the date/time marking the
boundary between academic years. Default is September 1st in user's locale}
}
\value{
A string e.g. "22/23"
}
\description{
Checks the current year and produces an academic year abbreviation, e.g.
"22/23" for the academic year spanning 2022 to 2023.
}
