% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submissions_marks.R
\name{get_all_quizzes}
\alias{get_all_quizzes}
\title{Get information about all quizzes in a module}
\usage{
get_all_quizzes(module_id)
}
\arguments{
\item{module_id}{A Canvas module ID number (NOT university module code).}
}
\value{
A tibble of information about all quizzes on the module.
}
\description{
Get information about all quizzes in a module
}
\examples{
\dontrun{cnvs::get_all_quizzes(module_id = 12345)}

}
