% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/announcements.R
\name{get_anncs}
\alias{get_anncs}
\title{Get list of pages for a module}
\usage{
get_anncs(module_id)
}
\arguments{
\item{module_id}{The Canvas ID of the module to update (use \code{\link[=get_module_id]{get_module_id()}})}
}
\description{
Get list of pages for a module
}
