% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_module_id}
\alias{get_module_id}
\title{Get Canvas module ID}
\usage{
get_module_id(search_term, academic_year)
}
\arguments{
\item{search_term}{A string containing word or phrase (exact) in the module
title to search for, OR a module code (not both!)}

\item{academic_year}{A string containing the academic year for the desired
module as e.g. "22/23"; defaults to the current year as calculated by
[cnvs::get_ac_year()]. Does not need to be specified for ongoing/non-academic
modules}
}
\value{
Canvas module ID as a numeric vector.
}
\description{
A nicer way to get the ID number for a module on Canvas. Essentially lets you
look up https://canvas.youruni.ac.uk/courses/THIS_NUMBER for use in other
Canvas API operations
}
