% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{get_module_list}
\alias{get_module_list}
\title{Get details of all modules}
\usage{
get_module_list(academic_modules = FALSE, academic_year = NA_character_)
}
\arguments{
\item{academic_modules}{Boolean. Return only academic modules (removing those
without a course ID code)? Defaults to `FALSE`.}

\item{academic_year}{Optional string to filter only specific academic
year(s). Accepts the format e.g. "22/23" or a vector of the same.}
}
\value{
A tibble of information about all the Canvas modules that are
  currently associated with the Canvas token set by [canvas_setup()].
}
\description{
Wrapper for `rcanvas` function [rcanvas::get_course_list()] with some extra
cleaning for convenience.
}
