% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pages.R
\name{get_pages}
\alias{get_pages}
\title{Get list of pages for a module}
\usage{
get_pages(module_id)
}
\arguments{
\item{module_id}{The Canvas ID of the module to update (use [get_module_id()])}
}
\description{
Get list of pages for a module
}
