% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quizzes.R
\name{get_quiz_groups}
\alias{get_quiz_groups}
\title{Get information about quiz groups}
\usage{
get_quiz_groups(module_id, quiz_id, quiz_group_id)
}
\arguments{
\item{module_id}{Canvas module ID}

\item{quiz_id}{Canvas quiz ID with existing quiz groups}

\item{quiz_group_id}{Canvas quiz IDs of existing quiz groups}
}
\value{
Tibble of quiz groups information
}
\description{
For a quiz with existing quiz groups, retrieve the info/settings about these
groups. Accepts one or multiple quiz group IDs but will ignore duplicate IDs.
}
