% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quizzes.R
\name{get_quiz_id}
\alias{get_quiz_id}
\title{Get a Canvas quiz ID for an existing quiz}
\usage{
get_quiz_id(module_id, search_term)
}
\arguments{
\item{module_id}{Canvas module ID number to search in (see \code{\link[=get_module_id]{get_module_id()}})}

\item{search_term}{String or regex to search the quiz names}
}
\value{
A vector of quiz IDs for all quizzes matching the search term
}
\description{
Note that this function will happily return more than one quiz ID, so if you
don't want that, do check your search term!
}
