% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submissions_marks.R
\name{get_quiz_info}
\alias{get_quiz_info}
\title{Get information about the quizzes on a module}
\usage{
get_quiz_info(module_id)
}
\arguments{
\item{module_id}{A Canvas module ID number (NOT university module code).}
}
\value{
A tibble of quiz information include both assignment and ID numbers.
Note that these are matched by the name/title of the quiz so THEY MUST BE
UNIQUE.
}
\description{
Get information about the quizzes on a module
}
