% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submissions_marks.R
\name{get_quiz_marks}
\alias{get_quiz_marks}
\title{Get a spreadsheet of full info and marks for all quizzes}
\usage{
get_quiz_marks(module_code, academic_year, quiz_title = "Worksheet")
}
\arguments{
\item{module_code}{A Sussex module code, e.g. "C8891".}

\item{academic_year}{A string containing the academic year for the desired
module as e.g. "22/23"; defaults to the current year as calculated by
\code{\link[=get_ac_year]{get_ac_year()}}. Does not need to be specified for
ongoing/non-academic modules}

\item{quiz_title}{Common element of the names of all quizzes to be used in
the calculation. Defaults to "Worksheet".}

\item{max_score}{Maximum possible score for the quizzes. Defaults to 7.}
}
\value{
A tibble of information about all submissions for all students for
all quizzes on a module
}
\description{
Get a spreadsheet of full info and marks for all quizzes
}
