% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submissions_marks.R
\name{get_quiz_questions}
\alias{get_quiz_questions}
\title{Get information about all questions in a quiz}
\usage{
get_quiz_questions(module_id, quiz_id)
}
\arguments{
\item{module_id}{A Canvas module ID number (NOT university module code).}

\item{quiz_id}{A Canvas quiz ID number}
}
\value{
A tibble of information about each question in the quiz.
}
\description{
Get information about all questions in a quiz
}
