% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_resp}
\alias{get_resp}
\title{Get a response from Canvas}
\usage{
get_resp(url, args)
}
\arguments{
\item{url}{A Canvas API URL}

\item{args}{Arguments to add to the query}
}
\value{
A tibble containing the requested information
}
\description{
Get a response from Canvas
}
