% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sections.R
\name{get_sections}
\alias{get_sections}
\title{Get all sections on a module}
\usage{
get_sections(module_id, search_term = "")
}
\arguments{
\item{module_id}{Canvas module ID number}

\item{search_term}{Optional string to search for in section names.}
}
\value{
Tibble of information about sections
}
\description{
Get all sections on a module
}
