% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submissions_marks.R
\name{get_submissions}
\alias{get_submissions}
\title{Get all submissions for a particular assignment}
\usage{
get_submissions(module_id, type, type_id, add_args = list())
}
\arguments{
\item{module_id}{A Canvas module ID number (NOT university module code).}

\item{type}{Either "quizzes" or "assignments".}

\item{type_id}{A Canvas assignment ID number.}

\item{add_args}{Any additional arguments to be added to the query}
}
\value{
A tibble of submission information for all submissions for all
  students.
}
\description{
Get all submissions for a particular assignment
}
\examples{
\dontrun{get_submissions("123C4", type = assignments, type_id = 123456)}
}
