% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submissions_marks.R
\name{get_submitted_responses}
\alias{get_submitted_responses}
\title{Get the responses submitted to a quiz}
\usage{
get_submitted_responses(module_id, type_id)
}
\arguments{
\item{module_id}{A Canvas module ID number (NOT university module code).}

\item{type_id}{A Canvas quiz ID number}
}
\value{
A long-form tibble of all submitted responses to all questions by everyone who submitted the quiz.
}
\description{
This has ONLY been tested on getting the responses to quiz questions, and will likely work different for retrieving submitted files!
}
