% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_users}
\alias{get_users}
\title{Get a tibble of all users on a module}
\usage{
get_users(module_id)
}
\arguments{
\item{module_id}{A Canvas module ID number (NOT university module code).}
}
\value{
A tibble of user information
}
\description{
Note that this is different from [cnvs::get_students()] as this gets all
USERS including convenors, observers, DTs etc
}
