% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quizzes.R
\name{make_args}
\alias{make_args}
\title{Convert a tibble to Canvas query arguments}
\usage{
make_args(data, type)
}
\arguments{
\item{data}{A tibble with variable names}

\item{type}{What type of arguments to generate. This string will be pasted
to the beginning of each argument name (except in the case of variable names
beginning with the string "answers"). For example, use "quiz" for use with
\code{\link[=create_quiz]{create_quiz()}} and "question" for \code{\link[=create_quiz_question]{create_quiz_question()}}.}
}
\value{
A list of arguments, one for each quiz question to be created.
}
\description{
This function converts a tibble of API parameters into a formatted args list
to pass on to an API query. The variable names MUST match allowable
parameters for the relevant query scope, and the values MUST match allowable
values (which I imagine is unsurprising).
}
