% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{next_weekday}
\alias{next_weekday}
\title{Calculate the date of the next weekday}
\usage{
next_weekday(date, weekday)
}
\arguments{
\item{date}{A datetime}

\item{weekday}{A day of the week as a number, with 1 = Sunday and 7 = Saturday}
}
\value{
The date of the next [weekday] after the given [date]
}
\description{
Taken from a [stack overflow solution by TimTeaFan](https://stackoverflow.com/questions/57893554/get-next-wednesday-date-after-a-date-with-r)
}
