% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{publish_module}
\alias{publish_module}
\title{Publish (offer) a Canvas module}
\usage{
publish_module(search_term)
}
\arguments{
\item{search_term}{Passed to \code{\link[=get_module_id]{get_module_id()}} to return the module ID}
}
\value{
Reply from Canvas
}
\description{
Publish (offer) a Canvas module
}
