% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/announcements.R
\name{quarto_annc}
\alias{quarto_annc}
\title{Post or update an announcement}
\usage{
quarto_annc(module_id, file_path, page_title, post_at = NULL)
}
\arguments{
\item{module_id}{The Canvas ID of the module to update (use
[cnvs::get_module_id()])}

\item{file_path}{Location of a .qmd Quarto file to post}

\item{page_title}{Optional. Custom page title, if different from the title of
the Quarto document}

\item{post_at}{Optional. Default is NULL which will post immediately. To
delay posting, provide a datetime in local time when the announcement
should be posted. Will be converted to UTC so Canvas doesn't mess it up.}
}
\value{
Response from Canvas
}
\description{
This function will create a new announcement if none currently exists with
the same title, and will update an existing announcement if the title is the
same. So, if you want announcements to be separated (and not overwritten), you
MUST NOT duplicate titles!
}
\examples{
\dontrun{quarto_annc("123C4", here::here("announcement.qmd"))}
}
