% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pages.R
\name{quarto_page}
\alias{quarto_page}
\title{Post or update a Canvas page from a Quarto document}
\usage{
quarto_page(
  module_id,
  file_path,
  syllabus = FALSE,
  page_id = "",
  page_title,
  overwrite = FALSE,
  published = TRUE,
  front_page = FALSE
)
}
\arguments{
\item{module_id}{The Canvas ID of the module to update (use \code{\link[=get_module_id]{get_module_id()}})}

\item{file_path}{Location of a .qmd Quarto file to post}

\item{syllabus}{Boolean. Whether the page to be uploaded is the module syllabus (\code{TRUE}) or not (\code{FALSE}, default)}

\item{page_id}{ID or URL of an existing Canvas page. If left as "" (the default) will instead create a new page}

\item{page_title}{Optional. Custom page title, if different from the title of the Quarto document}

\item{overwrite}{If there is an existing page already, should it be overwritten?}

\item{published}{Publish the page?}

\item{front_page}{Set as the module front page?}
}
\value{
Response from Canvas
}
\description{
Post or update a Canvas page from a Quarto document
}
