% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rcanvas_canvas_query}
\alias{rcanvas_canvas_query}
\title{Send a query to Canvas}
\usage{
rcanvas_canvas_query(urlx, args = NULL, type = "GET")
}
\arguments{
\item{urlx}{A Canvas API URL}

\item{args}{Arguments to append to the API request}

\item{type}{Type of query, of "GET", "POST", "PUT"}
}
\value{
Response from Canvas
}
\description{
Copied from the unexported `canvas_query()` function from [the {rcanvas} package](https://github.com/daranzolin/rcanvas).
See [cnvs::rcanvas_canvas_url()] for generating the API URL
}
