% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rcanvas_check_token}
\alias{rcanvas_check_token}
\title{Check Canvas token}
\usage{
rcanvas_check_token()
}
\value{
The user's Canvas token
}
\description{
Copied from the unexported `check_token()` function from [the {rcanvas} package](https://github.com/daranzolin/rcanvas).
}
