% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rcanvas_get_page}
\alias{rcanvas_get_page}
\title{Get a page from a response}
\usage{
rcanvas_get_page(resp, page)
}
\arguments{
\item{resp}{Response from Canvas}

\item{page}{Which page to get}
}
\value{
A URL
}
\description{
Copied from the unexported `get_page()` function from [the {rcanvas} package](https://github.com/daranzolin/rcanvas).
}
