% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rcanvas_has_rel}
\alias{rcanvas_has_rel}
\title{Check "rel" property (maybe?)}
\usage{
rcanvas_has_rel(x, rel)
}
\arguments{
\item{rel}{}
}
\value{
A Boolean
}
\description{
Copied from the unexported `has_rel()` function from [the {rcanvas} package](https://github.com/daranzolin/rcanvas).
Not 100% on what this one does but seems important!
}
