% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rcanvas_increment_pages}
\alias{rcanvas_increment_pages}
\title{Increment pages}
\usage{
rcanvas_increment_pages(base_url, n_pages)
}
\arguments{
\item{base_url}{A URL}

\item{n_pages}{Number of pages}
}
\value{
URL with incremented pages (probably)
}
\description{
Copied from the unexported `increment_pages()` function from [the {rcanvas} package](https://github.com/daranzolin/rcanvas).
}
