% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rcanvas_iter_args_list}
\alias{rcanvas_iter_args_list}
\title{Create a list of arguments for an API query}
\usage{
rcanvas_iter_args_list(x, label)
}
\arguments{
\item{x}{A vector of values}

\item{label}{Labels for the values of `x`}
}
\value{
A labeled (named) list
}
\description{
Copied from the unexported `iter_args_list()` function from [the {rcanvas} package](https://github.com/daranzolin/rcanvas).
}
