% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rcanvas_paginate}
\alias{rcanvas_paginate}
\title{Paginate a response from Canvas}
\usage{
rcanvas_paginate(x, showProgress = FALSE)
}
\arguments{
\item{x}{Response from Canvas from e.g. [cnvs::rcanvas_canvas_query()]}

\item{showProgress}{Show progress bar}
}
\value{
Paginated responses (probably)
}
\description{
Copied from the unexported `paginate()` function from [the {rcanvas} package](https://github.com/daranzolin/rcanvas).
}
