% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rcanvas_process_response}
\alias{rcanvas_process_response}
\title{Get and process a response from Canvas}
\usage{
rcanvas_process_response(url, args)
}
\arguments{
\item{url}{A Canvas API URL}

\item{args}{Arguments to append to the API request}
}
\value{
A dataset containing the requested information
}
\description{
Copied from the unexported `process_response()` function from [the {rcanvas} package](https://github.com/daranzolin/rcanvas).
}
