% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rcanvas_sc}
\alias{rcanvas_sc}
\title{Discard NULLs}
\usage{
rcanvas_sc(x)
}
\arguments{
\item{x}{A list or vector}
}
\value{
`x` with any NULL values removed
}
\description{
Copied from the unexported `sc()` function from [the {rcanvas} package](https://github.com/daranzolin/rcanvas).
}
