% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/announcements.R
\name{schedule_annc}
\alias{schedule_annc}
\title{Schedule an announcement from the document name}
\usage{
schedule_annc(module_id, annc_path = here::here("announcements"), annc_title)
}
\arguments{
\item{module_id}{The Canvas ID of the module to update (use
\code{\link[=get_module_id]{get_module_id()}})}

\item{annc_path}{Path to a folder containing announcements written in qmd files}

\item{annc_title}{Title to give the announcement on Canvas, if different than the title of the document}
}
\value{
Response from Canvas
}
\description{
This function assumes that you have a folder of announcement .qmd files whose
filenames are the date and time they should be posted in a format readable by
\code{\link[lubridate:ymd_hms]{lubridate::ymd_hm()}}, i.e. 202308310900 for 9:00 on August 31st 2023. This
function will find the next closest announcement to the current datetime and
post it to Canvas delayed until the time indicated in its filename.
}
\examples{
\dontrun{schedule_annc("123C4")}
}
