% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_module_id}
\alias{set_module_id}
\title{Check module code input and attempt to return the Canvas module ID}
\usage{
set_module_id(search_term, academic_year)
}
\arguments{
\item{search_term}{A string containing word or phrase (exact) in the module
title to search for, OR a module code (not both!)}

\item{academic_year}{A string containing the academic year for the desired
module as e.g. "22/23"; defaults to the current year as calculated by
[cnvs::get_ac_year()]. Does not need to be specified for ongoing/non-academic
modules}
}
\value{
Canvas module ID
}
\description{
Check module code input and attempt to return the Canvas module ID
}
