% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_renv_variables}
\alias{set_renv_variables}
\title{Set Up Canvas variables}
\usage{
set_renv_variables()
}
\value{
Sets .Renviron variables and runs \code{\link[=canvas_setup]{canvas_setup()}} to communicate with Canvas
}
\description{
Edits the user's .Renviron file with new or updated environ variables and sets these for use with Canvas/\code{cnvs}.
}
\details{
This interactive function guides the user through the creation (or updating, if they already exist) of two variables, CANVAS_TOKEN and CANVAS_DOMAIN. The function first checks if there are already variables defined in the local .Renviron file with these names, and gives the option to update them or not.
Once both values are put in, the function then sets these for use with Canvas using \code{\link[rcanvas:apihelpers]{rcanvas::set_canvas_token()}} and \code{\link[rcanvas:apihelpers]{rcanvas::set_canvas_domain()}} respectively.

To obtain a Canvas access token:
\itemize{
\item Log into Canvas as normal.
\item Click on your profile picture, then on Settings.
\item Click on the "+ New access token" button and follow the instructions.
}

Pro-tip: Save your Canvas token somewhere safe, where you can find it again. For example, paste it into a simple text document and save it as \code{canvas_token.txt} in a sensible place in your folders. Once you leave the page where it's displayed, you cannot retrieve it again!

To obtain the Canvas domain:
\itemize{
\item Navigate in your browser to your Canvas dashboard
\item The Canvas domain is everything after \verb{https://}, typically \code{canvas.your_uni_domain}, e.g. \code{canvas.sussex.ac.uk}
}
}
\examples{
 \dontrun{
 set_renv_variables()
}

}
