% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quizzes.R
\name{update_quiz}
\alias{update_quiz}
\title{Update existing Canvas quizzes}
\usage{
update_quiz(
  module_code,
  module_id,
  search_term,
  quiz_id,
  task = c("release_answers", "publish", "change_code", "update_description", "other"),
  description,
  other_args = list()
)
}
\arguments{
\item{module_code}{A string containing the module code in the format "123C4"
or "C1234". Either module_code or module_id must be provided.}

\item{module_id}{Canvas module ID number. Either module_code or module_id must be provided.}

\item{search_term}{Title or substring of the quiz name to modify. Either \code{search_term} or \code{quiz_id} must be provided.}

\item{quiz_id}{Canvas quiz ID number. Either \code{search_term} or \code{quiz_id} must be provided.}

\item{task}{What do you want to change about the quiz?
\itemize{
\item \code{release_answers}: Release both results and correct answers
\item \code{publish}: Publish an unpublished quiz
\item \code{change_code}: Change the access code for a quiz (to a random six-digit code)
\item \code{update_description}: Overwrite the description of the quiz.
\item \code{other}: Any other changes
}}

\item{description}{Optional - string containing the new description, if using "update_description".}

\item{other_args}{List of arguments to pass to Canvas when \code{task = "other"}. See
\href{https://canvas.instructure.com/doc/api/quizzes.html#method.quizzes/quizzes_api.create}{Canvas API documentation for quizzes}
for options.}
}
\value{
Response from Canvas.
}
\description{
Change settings for existing Canvas quizzes.
}
