% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{canvas_setup}
\alias{canvas_setup}
\title{Setup Canvas for API use}
\usage{
canvas_setup()
}
\value{
Sets the Canvas token and domain.
}
\description{
Wrapper for two \code{rcanvas} functions: \code{\link[rcanvas:apihelpers]{rcanvas::set_canvas_token()}} and
\code{\link[rcanvas:apihelpers]{rcanvas::set_canvas_domain()}}. Calls \code{\link[=Sys.getenv]{Sys.getenv()}} for both,
so this must be set up ahead of time (see Details).
}
\details{
To use this function, you must have a Canvas access token stored in your
Renviron file as \code{CANVAS_TOKEN}, and the domain URL for your institution
stored as \code{CANVAS_DOMAIN}. See \code{\link[=set_renv_variables]{set_renv_variables()}} for more.
}
