% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{cnvs}
\alias{cnvs}
\title{Canvas LMS API}
\usage{
cnvs(
  endpoint = "/api/v1/courses",
  ...,
  per_page = NULL,
  .token = NULL,
  .destfile = NULL,
  .overwrite = FALSE,
  .api_url = NULL,
  .method = "GET",
  .limit = NULL,
  .send_headers = NULL
)
}
\arguments{
\item{endpoint}{Canvas LMS API endpoint. Must be one of the following forms:

\itemize{
\item "METHOD path", e.g. "GET /api/v1/courses"
\item "path", e.g. "/api/v1/courses".
\item "METHOD url", e.g. "GET https://canvas.instructure.com/api/v1/courses"
\item "url", e.g. "https://canvas.instructure.com/api/v1/courses".
}

If the method is not supplied, will use \code{.method}, which defaults
to \code{GET}.}

\item{...}{Name-value pairs giving API parameters. Will be matched
into \code{url} placeholders, sent as query parameters in \code{GET}
requests, and in the JSON body of \code{POST} requests.}

\item{per_page}{Number of items to return per page. If omitted,
will be substituted by \code{max(.limit, 100)} if \code{.limit} is set,
otherwise determined by the API (never greater than 100).}

\item{.token}{Authentication token. Defaults to CANVAS_API_TOKEN
environment variable, if set.}

\item{.destfile}{path to write response to disk.  If NULL (default), response will
be processed and returned as an object.  If path is given, response will
be written to disk in the form sent.}

\item{.overwrite}{if \code{destfile} is provided, whether to overwrite an
existing file.  Defaults to FALSE.}

\item{.api_url}{Canvas domain. Used
if \code{endpoint} just contains a path. Defaults to CANVAS_DOMAIN
environment variable, if set.}

\item{.method}{HTTP method to use if not explicitly supplied in the
\code{endpoint}.}

\item{.limit}{Number of records to return. This can be used
instead of manual pagination. By default it is \code{NULL},
which means that the defaults of the Canvas API are used.
You can set it to a number to request more (or less)
records, and also to \code{Inf} to request all records.
Note, that if you request many records, then multiple Canva
API calls are used to get them, and this can take a potentially
long time.}

\item{.send_headers}{Named character vector of header field values
(excepting \code{Authorization}, which is handled via
\code{.token}). This can be used to override or augment the
defaults, which are as follows: the \code{Accept} field defaults
to \code{"application/json"}, the
\code{User-Agent} field defaults to
\code{"https://github.com/cwickham/cnvs"}, and the \code{Content-Type}
defaults to \code{"application/json"}.}
}
\value{
Answer from the API as a \code{cnvs_response} object, which is also a
\code{list}. Failed requests will generate an R error. Requests that
generate a raw response will return a raw vector.
}
\description{
Minimal wrapper to access the Canvas LMS API.

This is an extremely minimal client. You need to know the API
to be able to use this client. All this function does is:
\itemize{
\item Try to substitute each listed parameter into
\code{endpoint}, using the \code{:parameter} notation.
\item If a GET request (the default), then add
all other listed parameters as query parameters.
\item If not a GET request, then send the other parameters
in the request body, as JSON.
\item Convert the response to an R list using
\code{jsonlite::fromJSON}.
}
}
\examples{
\dontrun{
## Your courses
cnvs("/api/v1/courses")
}

}
\seealso{
\code{\link{cnvs_whoami}()} for details on Canvas API token
management.
}
