% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{cnvs_token}
\alias{cnvs_token}
\alias{cnvs_domain}
\title{Return the local user's Canvas token and domain}
\usage{
cnvs_token()

cnvs_domain()
}
\value{
A string, with the token, or a zero length string scalar,
if no token or domain is available.
}
\description{
You will need a Canvas access token, which you can get by following the instructions here:
\url{https://canvas.instructure.com/doc/api/file.oauth.html#manual-token-generation}
}
\details{
Canvas access tokens are specific to your user account on a specific
canvas domain, e.g. you cannot use the token for your University's site on
the instructure site (\url{https://canvas.instructure.com}).

Currently, it consults the \code{CANVAS_API_TOKEN} and \code{CANVAS_DOMAIN}
environment variables. Read more about setting these in \code{\link{cnvs_whoami}}.
}
