% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whoami.R
\name{cnvs_whoami}
\alias{cnvs_whoami}
\title{Info on current Canvas user, domain and token}
\usage{
cnvs_whoami(.token = NULL, .api_url = NULL, .send_headers = NULL)
}
\arguments{
\item{.token}{Authentication token. Defaults to CANVAS_API_TOKEN
environment variable, if set.}

\item{.api_url}{Canvas domain. Used
if \code{endpoint} just contains a path. Defaults to CANVAS_DOMAIN
environment variable, if set.}

\item{.send_headers}{Named character vector of header field values
(excepting \code{Authorization}, which is handled via
\code{.token}). This can be used to override or augment the
defaults, which are as follows: the \code{Accept} field defaults
to \code{"application/json"}, the
\code{User-Agent} field defaults to
\code{"https://github.com/cwickham/cnvs"}, and the \code{Content-Type}
defaults to \code{"application/json"}.}
}
\value{
A \code{cnvs_response} object, which is also a \code{list}.
}
\description{
Reports name and Canvas login ID for the current authenticated user, the
Canvas domain and the first bit of the token.
}
\details{
Your canvas domain is the address of your organizations' canvas site, e.g.
\code{https://oregonstate.instructure.com/}, provide this to \code{.api_url}.

Get a personal access token for your Canvas domain from
\code{\{CANVAS_DOMAIN/profile/settings\}}. The token itself is a string of 66
letters and digits. You can store it any way you like and provide explicitly
via the \code{.token} argument to \code{\link{cnvs}()}.

However, many prefer to define an environments variable \code{CANVAS_API_TOKEN},
and \code{CANVAS_DOMAIN}, with these values in their \code{.Renviron} file. Add
lines that looks like these, substituting your domain and token:

\preformatted{
CANVAS_DOMAIN="https://canvas.instructure.com"
CANVAS_API_TOKEN= "mvvGbKyGK9n5T57qhEu8K1sNMt85OLoNGTepqd3v5NEcWMuxArSz5aaXppPjodr5eU"
}

Put a line break at the end! If you’re using an editor that shows line
numbers, there should be (at least) three lines, where the third one is empty.
Restart R for this to take effect. Call \code{cmvs_whoami()} to confirm
success.

To get complete information on the authenticated user, call
\code{cnvs("/api/v1/user/self")}.
}
\examples{
\dontrun{
cnvs_whoami()

## explicit token + domain
cnvs_whoami(.token = "mvvGbKyGK9n5T57qhEu8K1sNMt85OLoNGTepqd3v5NEcWMuxArSz5aaXppPjodr5eU",
          .api_url = "https://canvas.instructure.com")
}
}
