% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whoami.R
\name{cnvs_whoami}
\alias{cnvs_whoami}
\title{Info on current Canvas user, domain and token}
\usage{
cnvs_whoami(.token = NULL, .api_url = NULL, .send_headers = NULL)
}
\arguments{
\item{.token}{Authentication token. Defaults to CANVAS_API_TOKEN
environment variable, if set.}

\item{.api_url}{Canvas domain. Used
if \code{endpoint} just contains a path. Defaults to CANVAS_DOMAIN
environment variable, if set.}

\item{.send_headers}{Named character vector of header field values
(excepting \code{Authorization}, which is handled via
\code{.token}). This can be used to override or augment the
defaults, which are as follows: the \code{Accept} field defaults
to \code{"application/json"} and the
\code{User-Agent} field defaults to
\code{"https://github.com/cwickham/cnvs"}. This can be used
to, e.g., provide a custom media type, in order to access a
preview feature of the API.}
}
\value{
A \code{cnvs_response} object, which is also a \code{list}.
}
\description{
Reports name and Canvas login ID for the current authenticated user, the
Canvas domain and the first bit of the token.
}
\examples{
\dontrun{
cnvs_whoami()

## explicit token + use with GitHub Enterprise
cnvs_whoami(.token = "mvvGbKyGK9n5T57qhEu8K1sNMt85OLoNGTepqd3v5NEcWMuxArSz5aaXppPjodr5eU",
          .api_url = "https://canvas.instructure.com")
}
}
