% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{create_link}
\alias{create_link}
\title{Create a markdown-style link for types of content}
\usage{
create_link(
  content,
  week_number,
  publish = FALSE,
  posit_link = FALSE,
  list_number,
  url = "https://analysingdata.pages.dev/content"
)
}
\arguments{
\item{content}{Type of module content}

\item{week_number}{Week of term}

\item{publish}{Boolean. If TRUE the text will be linked, otherwise will just return formatted text with no link}

\item{posit_link}{Whether the link should go to the Posit workspace. Calls \code{\link[=create_posit_link]{create_posit_link()}}}

\item{list_number}{The numerical ID of the Posit list. Only needed if \code{posit_link = TRUE}.}

\item{url}{Stem URL to build on for non-Posit links}
}
\value{
A text string
}
\description{
Helper function to build in-text links for use on Canvas pages.
Specifically designed to work with our module website and reference docs.
}
\details{
Assumes that the final (non-Posit) URL will be in the format:
\code{url}/\code{content}s/\code{content}_\code{week} (with \code{week} given a leading 0 if needed)
So if \code{content} is \code{skills_lab} and the \code{week} is 5, the resulting URL will be:
\code{url}/skills_labs/skills_lab_05
Look I'm sorry if that's not how you roll, this function works for me!
}
