% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quizzes.R
\name{create_quiz_info}
\alias{create_quiz_info}
\title{Create quiz metadata}
\usage{
create_quiz_info(
  module_code,
  quiz_desc = NULL,
  first_unlock_date,
  unlock_time,
  last_lock_date,
  lock_time,
  mopup_day,
  quiz_type = c("assignment", "practice_quiz", "graded_survey", "survey"),
  make_practice = 1,
  quiz_name_fn = function(x) paste("Week", x, "Worksheet", collapse = " "),
  n_questions = 5,
  standard_time_limit = 12,
  has_break_weeks = TRUE,
  break_weeks = 11:13,
  challengr = FALSE
)
}
\arguments{
\item{module_code}{string containing the module code in the format "123C4"
or "C1234".}

\item{quiz_desc}{Quiz description as a string, which can include basic HTML
formatting tags. If `NULL` (the default), runs
[cnvs::create_quiz_desc()] with default settings. Also accepts
either a single string (which will be applied to all quizzes) or a list of
two descriptions, named `practice` and `marked`. The `practice` description
will only be used if `first_quiz_practice` is `TRUE` and will be used for
the first quiz only.}

\item{first_unlock_date}{Date as "YYYY/MM/DD" of the date that the first quiz
of term should unlock.}

\item{unlock_time}{Hours as a number past midnight when the quizzes should
unlock each week. E.g. for 9am, enter `9`.}

\item{last_lock_date}{Date as "YYYY/MM/DD" of the date that the last quiz of
term should lock.}

\item{lock_time}{Hours as a number past midnight when the quizzes should lock
each week. E.g. for 6pm, enter `18`.}

\item{mopup_day}{Optional. If mop-up quizzes are allowed, what day should the quiz close on? Lock time will default to 11:59 on that day.}

\item{quiz_type}{Either one of "assignment", "practice_quiz",
"graded_survey", or "survey" (which will be applied to all) or a vector of
combinations of the same in the order in which they should be applied.}

\item{make_practice}{Which quiz(zes) should be practice quizzes, rather than
marked? Defaults to `1` but accepts a numeric vector corresponding to week
numbers, or `NULL` if none should be changed. For those quizzes indicated,
the following will be changed:
  * `quiz_type` will be set to "practice_quiz"
  * `quiz_name` will be updated so the word "(Practice)" is at the end of
  the title
  * `quiz_desc` will be updated to the description stored in "$practice"
  * `allowed_attempts` will be set to -1 (which will translate to
  unlimited)
  * `access_code` will be set to FALSE (which will be converted to NULL
  when combined with [convenrhelpr::create_quizzes()])}

\item{quiz_name_fn}{A function to name the quizzes. Default is `function(x)
paste("Week", x, "Worksheet", collapse = " ")`. Minimally requires `x`,
which is a number corresponding to the week of term for that quiz.}

\item{n_questions}{Total number of questions the quiz will contain}

\item{standard_time_limit}{"Standard" amount of time for the quiz. Actual
quiz length will be 1.5 times longer (for 50% extra time)}

\item{has_break_weeks}{Logical. Does the module you are making quizzes for
have break weeks (e.g. spring break)?}

\item{break_weeks}{Numerical vector indicating weeks out of the total number
of weeks that the term spans are break weeks (or otherwise don't have
quizzes). E.g. for an 11-week term, with three weeks of spring break, the
total number of weeks is 14: 10 weeks of term, three weeks break, and one
final week of term. In this case the input would be 11:13.}

\item{challengr}{}
}
\value{
A tibble of quiz information to pass to
  [convenrhelpr::create_quizzes()].
}
\description{
Make a tibble of information to use in [cnvs::create_quizzes()].
Specifically designed to make weekly quizzes for worksheets.
}
