% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quizzes.R
\name{create_quizzes}
\alias{create_quizzes}
\title{Create Canvas quizzes}
\usage{
create_quizzes(
  week,
  quiz_name,
  unlock,
  lock,
  module_id,
  quiz_desc,
  quiz_type,
  time_limit,
  allowed_attempts,
  access_code
)
}
\arguments{
\item{quiz_name}{Title of the quiz to be created}

\item{unlock}{A datetime using [format_ISO8601()] indicating when the quiz
should be unlocked.}

\item{lock}{A datetime using [format_ISO8601()] indicating when the quiz
should be locked.}

\item{module_id}{Canvas ID number of the module where the quizzes will be created
(note: NOT module code!)}

\item{quiz_desc}{Single string containing the text description for the quiz.
Accepts basic HTML formatting tags.}

\item{time_limit}{Time limit for the quiz in minutes.}

\item{allowed_attempts}{A number, with -1 corresponded to infinite attempts.}

\item{access_code}{Logical. If TRUE, a random 6-digit code will be generated.
If FALSE, no code will be required to access the quiz.}
}
\value{
Answer from Canvas; quizzes should appear on Canvas.
}
\description{
Generate weekly worksheet quizzes for your Canvas site. For making multiple
quizzes at once, use in conjunction with [purrr::pmap()] and a tibble of
associated quiz info generated by [cnvs::create_quiz_info()].
}
\seealso{
[cnvs::get_module_id()]
}
