% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{get_students}
\alias{get_students}
\title{Get information about all students on a module}
\usage{
get_students(module_id)
}
\arguments{
\item{module_id}{A Canvas module ID number (NOT university module code).}
}
\value{
A tibble of student information
}
\description{
Get information about all students on a module
}
