% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rcanvas_canvas_url}
\alias{rcanvas_canvas_url}
\title{Generate a Canvas URL}
\usage{
rcanvas_canvas_url()
}
\value{
A Canvas API URL
}
\description{
Copied from the unexported `canvas_url()` function from [the {rcanvas} package](https://github.com/daranzolin/rcanvas).
}
